Refactor: add `commandExists` function
